+++ src/cast_all_rays.c
@@ -6,13 +6,13 @@
 /*   By: joeduard <joeduard@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 16:44:17 by azamario          #+#    #+#             */
-/*   Updated: 2022/11/25 17:05:06 by joeduard         ###   ########.fr       */
+/*   Updated: 2022/11/26 00:29:26 by joeduard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/cub3d.h"
 
-float	normalize_angle_cast_ray(float angle)
+double	normalize_angle_cast_ray(double angle)
 {
 	angle = remainder(angle, TWO_PI);
 	if (angle < 0)
@@ -20,15 +20,15 @@ float	normalize_angle_cast_ray(float angle)
 	return (angle);
 }
 
-float	distance_between_points(float x1, float y1, float x2, float y2)
+double	distance_between_points(double x1, double y1, double x2, double y2)
 {
-	float	result;
+	double	result;
 
 	result = sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
 	return (result);
 }
 
-void	where_is_ray_facing(float ray_angle, t_game *game)
+void	where_is_ray_facing(double ray_angle, t_game *game)
 {
 	game->rays->is_ray_facing_down = ray_angle > 0 && ray_angle < PI;
 	game->rays->is_ray_facing_up = !game->rays->is_ray_facing_down;
@@ -37,7 +37,7 @@ void	where_is_ray_facing(float ray_angle, t_game *game)
 	game->rays->is_ray_facing_left = !game->rays->is_ray_facing_right;
 }
 
-void	cast_ray(float ray_angle, int strip_id, t_game *game)
+void	cast_ray(double ray_angle, int strip_id, t_game *game)
 {
 	ray_angle = normalize_angle_cast_ray(ray_angle);
 	where_is_ray_facing(ray_angle, game);
@@ -49,7 +49,7 @@ void	cast_ray(float ray_angle, int strip_id, t_game *game)
 //start first ray subtracting half of our FOV
 void	cast_all_rays(t_game *game)
 {
-	float	ray_angle;
+	double	ray_angle;
 	int		strip_id;
 
 	ray_angle = game->player.rotation_angle - (FOV_ANGLE / 2);
